+++ lib/nnet-component.cpp
@@ -6,7 +6,7 @@
 
 namespace snowboy {
 
-	int32_t ChunkInfo::GetIndex(int32_t offset) const {
+	size_t ChunkInfo::GetIndex(size_t offset) const {
 		if (m_offsets.empty())
 		{ // if data is contiguous
 			SNOWBOY_ASSERT((offset <= m_last_offset) && (offset >= m_first_offset));
@@ -16,14 +16,14 @@ namespace snowboy {
 			auto iter = std::lower_bound(m_offsets.begin(), m_offsets.end(), offset);
 			// make sure offset is present in the vector
 			SNOWBOY_ASSERT(iter != m_offsets.end() && *iter == offset);
-			return static_cast<int32_t>(iter - m_offsets.begin());
+			return static_cast<size_t>(iter - m_offsets.begin());
 		}
 	}
 
-	int32_t ChunkInfo::GetOffset(int32_t index) const {
+	size_t ChunkInfo::GetOffset(size_t index) const {
 		if (m_offsets.empty())
 		{											 // if data is contiguous
-			int32_t offset = index + m_first_offset; // just offset by the first_offset_
+			size_t offset = index + m_first_offset; // just offset by the first_offset_
 			SNOWBOY_ASSERT((offset <= m_last_offset) && (offset >= m_first_offset));
 			return offset;
 		} else
@@ -510,7 +510,7 @@ namespace snowboy {
 				// but is restricted to chunk 0 for efficiency reasons
 				for (size_t c = 0; c < num_splice; c++)
 				{
-					for (int32_t out_index = 0; out_index < out_chunk_size; out_index++)
+					for (size_t out_index = 0; out_index < out_chunk_size; out_index++)
 					{
 						int32_t out_offset = out_info.GetOffset(out_index);
 						int32_t in_index = in_info.GetIndex(out_offset + m_context[c]);
@@ -522,7 +522,7 @@ namespace snowboy {
 				// and offset these by input chunk size
 				for (size_t c = 0; c < num_splice; c++)
 				{
-					for (int32_t out_index = 0; out_index < out_chunk_size; out_index++)
+					for (size_t out_index = 0; out_index < out_chunk_size; out_index++)
 					{
 						int32_t last_value = indexes[c][(chunk - 1) * out_chunk_size + out_index];
 						indexes[c][chunk * out_chunk_size + out_index] = (last_value == -1 ? -1 : last_value + in_chunk_size);
@@ -531,7 +531,7 @@ namespace snowboy {
 			}
 			if (const_dim != 0)
 			{
-				for (int32_t out_index = 0; out_index < out_chunk_size; out_index++)
+				for (size_t out_index = 0; out_index < out_chunk_size; out_index++)
 					const_indexes[chunk * out_chunk_size + out_index] = chunk * in_chunk_size + out_index; // there is
 																										   // an arbitrariness here; since we assume the const_component
 																										   // is constant within a chunk, it doesn't matter from where we copy.
@@ -540,7 +540,7 @@ namespace snowboy {
 
 		for (size_t c = 0; c < num_splice; c++)
 		{
-			int32_t dim = input_dim - const_dim; // dimension we
+			size_t dim = input_dim - const_dim; // dimension we
 			// are splicing
 			SubMatrix in_part(in, 0, in.m_rows, 0, dim);
 			SubMatrix out_part(*out, 0, out->m_rows, c * dim, dim);
